requirements.txt
absl-py==2.1.0
accelerate==0.34.2
biopandas==0.5.1
bitsandbytes==0.44.1
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cosine_annealing_warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
docopt==0.6.2
et-xmlfile==1.1.0
fastavro==1.9.7
filelock==3.16.1
freesasa==2.2.1
fsspec==2024.9.0
ftpretty==0.4.0
goatools==1.4.12
google-auth==2.35.0
google-auth-oauthlib==1.0.0
grpcio==1.66.2
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver
looseversion==1.1.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib 
mdurl==0.1.2
mkl-fft 
mkl-random
mkl-service==2.4.0
mmtf-python==1.1.3
mpmath==1.3.0
msgpack==1.1.0
networkx==3.1
numexpr 
numpy 
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openpyxl==3.1.5
pandas 
patsy==0.5.6
peft==0.13.0
ply==3.11
prot2token==0.1.0
proteinshake==0.3.14
protobuf==5.28.2
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydot==3.0.2
Pygments==2.18.0
pyproject_hooks==1.2.0
PyQt5==5.15.10
python-box==7.2.0
PyYAML==6.0.2
rdkit-pypi==2022.9.5
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
statsmodels==0.14.1
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==1.0.9
tokenizers==0.20.0
tomli @ 
torch==2.4.1
torchvision==0.19.1
tornado 
tqdm==4.66.5
transformers==4.45.1
triton==3.0.0
typing_extensions==4.12.2
tzdata 
unicodedata2
urllib3==2.2.3
Werkzeug==3.0.4
XlsxWriter==3.2.0
zipp==3.20.2
